+++ inc/hud.h
@@ -6,9 +6,9 @@
 #include "hud_consts.h"
 
 typedef struct {
-    u8 hundrs;
-    u8 tens;
-    u8 ones;
+    u16 hundrs;
+    u16 tens;
+    u16 ones;
 } Digits;
 
 u16 hud_loadInitialState (u16 currentTileIndex);
@@ -17,26 +17,26 @@ void hud_free_dst_buffer ();
 void hud_free_pals_buffer ();
 
 void hud_resetAmmo ();
-void hud_setAmmo (u8 hundreds, u8 tens, u8 ones);
+void hud_setAmmo (u16 hundreds, u16 tens, u16 ones);
 void hud_addAmmoUnits (u16 amnt);
 void hud_subAmmoUnits (u16 amnt);
 void hud_resetHealth ();
-void hud_setHealth (u8 hundreds, u8 tens, u8 ones);
+void hud_setHealth (u16 hundreds, u16 tens, u16 ones);
 void hud_addHealthUnits (u16 amnt);
 void hud_subHealthUnits (u16 amnt);
 void hud_resetArmor ();
-void hud_setArmor (u8 hundreds, u8 tens, u8 ones);
+void hud_setArmor (u16 hundreds, u16 tens, u16 ones);
 void hud_addArmorUnits (u16 amnt);
 void hud_subArmorUnits (u16 amnt);
 void hud_resetWeapons ();
-void hud_addWeapon (u8 weapon);
+void hud_addWeapon (u16 weapon);
 void hud_resetKeys ();
-void hud_addKey (u8 key);
+void hud_addKey (u16 key);
 void hud_resetFaceExpression ();
-void hud_setFaceExpression (u8 expr, u8 timer);
+void hud_setFaceExpression (u16 expr, u16 timer);
 
-bool hud_hasWeaponInInventory (u8 weapon);
-bool hud_hasKeyInInventory (u8 key);
+bool hud_hasWeaponInInventory (u16 weapon);
+bool hud_hasKeyInInventory (u16 key);
 bool hud_isDead ();
 
 void hud_update ();
